requirements.txt
pandas>=1.3.0
numpy>=1.20.0
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.18.0
Pillow>=8.0.0
requests>=2.26.0
tqdm>=4.62.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
regex>=2021.8.3 